+++ relay-server/src/metrics_extraction/transactions.rs
@@ -1,19 +1,17 @@
+use crate::metrics_extraction::conditional_tagging::run_conditional_tagging;
+use crate::metrics_extraction::utils;
+use crate::metrics_extraction::TaggingRule;
+use relay_common::{SpanStatus, UnixTimestamp};
+use relay_general::protocol::TraceContext;
+use relay_general::protocol::{AsPair, Timestamp, TransactionSource};
+use relay_general::protocol::{Context, ContextInner};
+use relay_general::protocol::{Event, EventType};
+use relay_general::store;
+use relay_general::types::Annotated;
+use relay_metrics::{DurationUnit, Metric, MetricNamespace, MetricUnit, MetricValue};
 use serde::{Deserialize, Serialize};
 use std::collections::{BTreeMap, BTreeSet};
-
-#[cfg(feature = "processing")]
-use {
-    crate::metrics_extraction::conditional_tagging::run_conditional_tagging,
-    crate::metrics_extraction::{utils, TaggingRule},
-    relay_common::{SpanStatus, UnixTimestamp},
-    relay_general::protocol::TraceContext,
-    relay_general::protocol::{AsPair, Event, EventType, Timestamp, TransactionSource},
-    relay_general::protocol::{Context, ContextInner},
-    relay_general::store,
-    relay_general::types::Annotated,
-    relay_metrics::{DurationUnit, Metric, MetricNamespace, MetricUnit, MetricValue},
-    std::fmt,
-};
+use std::fmt;
 
 /// The metric on which the user satisfaction threshold is applied.
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -51,20 +49,31 @@ pub struct CustomMeasurementConfig {
     limit: usize,
 }
 
+/// Maximum supported version of metrics extraction from transactions.
+///
+/// The version is an integer scalar, incremented by one on each new version.
+const EXTRACT_MAX_VERSION: u16 = 1;
+
 /// Configuration for extracting metrics from transaction payloads.
 #[derive(Default, Debug, Clone, Serialize, Deserialize)]
 #[serde(default, rename_all = "camelCase")]
 pub struct TransactionMetricsConfig {
+    /// The required version to extract transaction metrics.
+    version: u16,
     extract_metrics: BTreeSet<String>,
     extract_custom_tags: BTreeSet<String>,
     satisfaction_thresholds: Option<SatisfactionConfig>,
     custom_measurements: CustomMeasurementConfig,
 }
 
-#[cfg(feature = "processing")]
+impl TransactionMetricsConfig {
+    pub fn is_enabled(&self) -> bool {
+        self.version > 0 && self.version <= EXTRACT_MAX_VERSION
+    }
+}
+
 const METRIC_NAMESPACE: MetricNamespace = MetricNamespace::Transactions;
 
-#[cfg(feature = "processing")]
 fn get_trace_context(event: &Event) -> Option<&TraceContext> {
     let contexts = event.contexts.value()?;
     let trace = contexts.get("trace").map(Annotated::value);
@@ -77,18 +86,15 @@ fn get_trace_context(event: &Event) -> Option<&TraceContext> {
 
 /// Extract transaction status, defaulting to [`SpanStatus::Unknown`].
 /// Must be consistent with `process_trace_context` in [`relay_general::store`].
-#[cfg(feature = "processing")]
 fn extract_transaction_status(trace_context: &TraceContext) -> SpanStatus {
     *trace_context.status.value().unwrap_or(&SpanStatus::Unknown)
 }
 
-#[cfg(feature = "processing")]
 fn extract_transaction_op(trace_context: &TraceContext) -> Option<String> {
     let op = trace_context.op.value()?;
     Some(op.to_string())
 }
 
-#[cfg(feature = "processing")]
 fn extract_dist(transaction: &Event) -> Option<String> {
     let mut dist = transaction.dist.0.clone();
     store::normalize_dist(&mut dist);
@@ -97,7 +103,6 @@ fn extract_dist(transaction: &Event) -> Option<String> {
 
 /// Extract HTTP method
 /// See <https://github.com/getsentry/snuba/blob/2e038c13a50735d58cc9397a29155ab5422a62e5/snuba/datasets/errors_processor.py#L64-L67>.
-#[cfg(feature = "processing")]
 fn extract_http_method(transaction: &Event) -> Option<String> {
     let request = transaction.request.value()?;
     let method = request.method.value()?;
@@ -106,14 +111,12 @@ fn extract_http_method(transaction: &Event) -> Option<String> {
 
 /// Satisfaction value used for Apdex and User Misery
 /// <https://docs.sentry.io/product/performance/metrics/#apdex>
-#[cfg(feature = "processing")]
 enum UserSatisfaction {
     Satisfied,
     Tolerated,
     Frustrated,
 }
 
-#[cfg(feature = "processing")]
 impl fmt::Display for UserSatisfaction {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
@@ -124,7 +127,6 @@ impl fmt::Display for UserSatisfaction {
     }
 }
 
-#[cfg(feature = "processing")]
 impl UserSatisfaction {
     /// The frustration threshold is always four times the threshold
     /// (see <https://docs.sentry.io/product/performance/metrics/#apdex>)
@@ -143,7 +145,6 @@ impl UserSatisfaction {
 
 /// Extract the the satisfaction value depending on the actual measurement/duration value
 /// and the configured threshold.
-#[cfg(feature = "processing")]
 fn extract_user_satisfaction(
     config: &Option<SatisfactionConfig>,
     transaction: &Event,
@@ -172,7 +173,6 @@ fn extract_user_satisfaction(
 /// Decide whether we want to keep the transaction name.
 /// High-cardinality sources are excluded to protect our metrics infrastructure.
 /// Note that this will produce a discrepancy between metrics and raw transaction data.
-#[cfg(feature = "processing")]
 fn keep_transaction_name(source: &TransactionSource) -> bool {
     match source {
         // For now, we hope that custom transaction names set by users are low-cardinality.
@@ -200,7 +200,6 @@ fn keep_transaction_name(source: &TransactionSource) -> bool {
 }
 
 /// These are the tags that are added to all extracted metrics.
-#[cfg(feature = "processing")]
 fn extract_universal_tags(
     event: &Event,
     custom_tags: &BTreeSet<String>,
@@ -265,7 +264,6 @@ fn extract_universal_tags(
 ///
 /// For known measurements, this returns `Some(MetricUnit)`, which can also include
 /// `Some(MetricUnit::None)`. For unknown measurement names, this returns `None`.
-#[cfg(feature = "processing")]
 fn get_metric_measurement_unit(metric: &str) -> Option<MetricUnit> {
     match metric {
         // Web
@@ -294,7 +292,6 @@ fn get_metric_measurement_unit(metric: &str) -> Option<MetricUnit> {
     }
 }
 
-#[cfg(feature = "processing")]
 pub fn extract_transaction_metrics(
     config: &TransactionMetricsConfig,
     breakdowns_config: Option<&store::BreakdownsConfig>,
@@ -334,7 +331,6 @@ pub fn extract_transaction_metrics(
     !added_slice.is_empty()
 }
 
-#[cfg(feature = "processing")]
 fn extract_transaction_metrics_inner(
     config: &TransactionMetricsConfig,
     breakdowns_config: Option<&store::BreakdownsConfig>,
@@ -468,7 +464,6 @@ fn extract_transaction_metrics_inner(
     }
 }
 
-#[cfg(feature = "processing")]
 fn get_measurement_rating(name: &str, value: f64) -> Option<String> {
     let rate_range = |meh_ceiling: f64, poor_ceiling: f64| {
         debug_assert!(meh_ceiling < poor_ceiling);