+++ relay-server/src/statsd.rs
@@ -309,7 +309,6 @@ pub enum RelayTimers {
     OutcomeAggregatorFlushTime,
 
     /// Time in milliseconds spent on converting a transaction event into a metric.
-    #[cfg(feature = "processing")]
     TransactionMetricsExtraction,
 }
 
@@ -338,7 +337,6 @@ impl TimerMetric for RelayTimers {
             RelayTimers::UpstreamRequestsDuration => "upstream.requests.duration",
             RelayTimers::TimestampDelay => "requests.timestamp_delay",
             RelayTimers::OutcomeAggregatorFlushTime => "outcomes.aggregator.flush_time",
-            #[cfg(feature = "processing")]
             RelayTimers::TransactionMetricsExtraction => "metrics.extraction.transactions",
         }
     }